buildserver: install default SDK packages using fdroid/sdkmanager
fdroid/sdkmanager provides a root of trust to verify all the packages it
downloads, so it fully replaces what makebuildserver was doing.

This also upgrades to the latest version of Android SDK "tools".  We
have been frozen on tools_r25.2.5-linux.zip for the longest time
because the _buildserver_ provisioning required `android update sdk`
to work.  This replaces that with fdroid/sdkmanager, so the
_buildserver_ is free to have any version of "tools" installed, since
none of the _buildserver_'s setup processes need anything from "tools"
any more.